moved selectdemo to components, settings almost done
